pyproject.toml
[build-system]
requires = ["setuptools >= 66", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# this is the name of the package: name that will appear if one `pip list` or `pip install open-darts`
name = "open-darts"
# This version has to be incremented manually for every new release.
version = "1.3.1"
authors = [
    {name = "Denis Voskov", email = "D.V.Voskov@tudelft.nl"}
]
maintainers = [
    {name = "Ilshat Saifullin", email = "i.s.saifullin@tudelft.nl"}
]

description = "Open Delft Advanced Research Terra Simulator"
readme = "README.md"
keywords = ["energy transition", "modeling of CO2 sequestration", "geothermal energy production"]
requires-python = ">=3.9"
# Here we specify open-darts dependencies, at the moment we do not specify versions of each one of those, but that can be the case.
dependencies = ['numba', 'scipy', 'pandas', 'meshio', 'gmsh', 'iapws', 'openpyxl', 'pyevtk', 'matplotlib', 'vtk',
    'shapely', 'igraph', 'sympy', 'opmcpg', 'xarray', 'netCDF4', 'h5py', 'open-darts-flash', 'phreeqpy']

license = {text = "LICENSE.md"}

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: C++"
]
dynamic = ["scripts"]

[project.urls]
homepage = "https://darts.citg.tudelft.nl/"
repository = "https://gitlab.com/open-darts/open-darts"
documentation = "https://open-darts.readthedocs.io/en/docs"

[project.optional-dependencies]
# These are groups of optional dependencies: not all users need ot install the dependencies. They are specific for a given fucntionality.
# to install darts with this dependencies: `pip install .[docs]`

# Dependencies to be able to build the documentation
docs = ["myst-parser", "sphinx_rtd_theme", "sphinx", "sphinx-tabs", "sphinx_inline_tabs", "sphinxcontrib-matlabdomain", "sphinxcontrib-bibtex", "linkify-it-py", "sphinx-copybutton"]
